requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
numpy
scipy
streamlit
matplotlib
python-dotenv
pytest
httpx
plotly
pandas